moving cmd inside pkg dir...
